requirements.txt
semantic-kernel
snowflake-connector-python
pyautogen
sympy